Refactored utils file.
